+++ src/DependencyInjection/Configuration.php
@@ -14,7 +14,6 @@ class Configuration implements ConfigurationInterface
     public const ROOT_NODE = 'dbp_relay_base_organization_connector_campusonline';
     public const DATABASE_URL = 'database_url';
     public const CAMPUS_ONLINE_NODE = 'campus_online';
-    public const LEGACY_NODE = 'legacy';
     public const BASE_URL_NODE = 'base_url';
     public const CLIENT_ID_NODE = 'client_id';
     public const CLIENT_SECRET_NODE = 'client_secret';
@@ -31,12 +30,9 @@ class Configuration implements ConfigurationInterface
                 ->end()
                 ->arrayNode(self::CAMPUS_ONLINE_NODE)
                     ->children()
-                        ->booleanNode(self::LEGACY_NODE)
-                            ->defaultTrue()
-                        ->end()
-                            ->scalarNode(self::BASE_URL_NODE)->end()
-                            ->scalarNode(self::CLIENT_ID_NODE)->end()
-                            ->scalarNode(self::CLIENT_SECRET_NODE)->end()
+                        ->scalarNode(self::BASE_URL_NODE)->end()
+                        ->scalarNode(self::CLIENT_ID_NODE)->end()
+                        ->scalarNode(self::CLIENT_SECRET_NODE)->end()
                     ->end()
                 ->end()
                 ->scalarNode(self::CACHE_REFRESH_INTERVAL_NODE)